+++ Sources/Core/BoxSDKConfiguration.swift
@@ -59,6 +59,10 @@ public struct BoxSDKConfiguration {
     public let consoleLogDestination: ConsoleLogDestination
     /// File log destination.
     public let fileLogDestination: FileLogDestination?
+    /// An optional custom callback URL string. The URL to which Box redirects the browser when authentication completes.
+    /// The user's actual interaction with your application begins when Box redirects to this URL.
+    /// If not specified, default URL is used in a format of `boxsdk-clientId://boxsdkoauth2redirect` with the real value of `clientId`.
+    public let callbackURL: String
 
     /// Initializer
     ///
@@ -73,6 +77,7 @@ public struct BoxSDKConfiguration {
     ///   - consoleLogDestination: Console log destination.
     ///   - fileLogDestination: File log destination.
     ///   - clientAnalyticsInfo: Analytics info that is set to request headers.
+    ///   - callbackURL: Optional callback URL for OAuth 2 authentication flow
     init(
         clientId: String = "",
         clientSecret: String = "",
@@ -84,7 +89,8 @@ public struct BoxSDKConfiguration {
         retryBaseInterval: TimeInterval? = defaultRetryBaseInterval,
         consoleLogDestination: ConsoleLogDestination? = ConsoleLogDestination(),
         fileLogDestination: FileLogDestination? = nil,
-        clientAnalyticsInfo: ClientAnalyticsInfo? = nil
+        clientAnalyticsInfo: ClientAnalyticsInfo? = nil,
+        callbackURL: String? = nil
     ) throws {
         self.clientId = clientId
         self.clientSecret = clientSecret
@@ -92,6 +98,9 @@ public struct BoxSDKConfiguration {
         self.apiBaseURL = apiBaseURL ?? defaultAPIBaseURL
         self.uploadApiBaseURL = uploadApiBaseURL ?? defaultUploadAPIBaseURL
         self.oauth2AuthorizeURL = oauth2AuthorizeURL ?? defaultOAuth2AuthorizeURL
+
+        let defaultCallbackURL = "boxsdk-\(clientId)://boxsdkoauth2redirect"
+        self.callbackURL = callbackURL ?? defaultCallbackURL
 
         try URLValidation.validate(networkUrl: self.apiBaseURL)
         try URLValidation.validate(networkUrl: self.uploadApiBaseURL)